requirements.txt
numpy
matplotlib
pillow
logging
toml

